v0.3.0

+++ cartographer/include/TypeOne.h
@@ -6,6 +6,6 @@ using namespace std;
 namespace TypeOne
 {
     void Process(std::string &key, std::string &value);
-    std::set<std::set<int>> Get();
-    extern std::map<int, uint64_t> blockCount;
+    std::set<std::set<int64_t>> Get();
+    extern std::map<int64_t, uint64_t> blockCount;
 } // namespace TypeOne
